test(Task 3 Story 2): As a PO, I want the Terminal to read the 'number of items' and to display it; Task read from console for real

+++ src/test/java/com/orange/dojo/elephantcarpaccio/TerminalReadInputsTest.java
@@ -2,7 +2,7 @@ package com.orange.dojo.elephantcarpaccio;
 
 import org.junit.Test;
 
-import java.util.OptionalInt;
+import java.util.OptionalDouble;
 
 import static org.fest.assertions.Assertions.assertThat;
 
@@ -16,7 +16,7 @@ public class TerminalReadInputsTest {
 
     // when
     NumberOfItems numberOfItems = terminal.readNumberOfItems(
-            new InputReader("Number of Items", OptionalInt.of(testModeExpectedInputValue)));
+            new InputReader("Number of Items", OptionalDouble.of(testModeExpectedInputValue)));
 
     // then
     assertThat(numberOfItems).isEqualTo(new NumberOfItems(testModeExpectedInputValue));
